Moved the provider to a subdirectory
