+++ concourse/data_caller_identity.go
@@ -1,4 +1,4 @@
-package main
+package concourse
 
 import (
 	"fmt"
