+++ concourse/data_server_info.go
@@ -1,4 +1,4 @@
-package main
+package concourse
 
 import (
 	"fmt"
