+++ concourse/resource_team.go
@@ -1,4 +1,4 @@
-package main
+package concourse
 
 import (
 		"fmt"
